refactor(cli)!: promote `self update` to top-level `update`
The `self` parent held exactly one subcommand, so it added a level of nesting
without grouping anything. Moving update to the top level makes it reachable as
`devsy update`.

BREAKING CHANGE: `devsy self update` is no longer recognized. Use `devsy update`.
