--- cmd/self/self.go
@@ -1,16 +1,0 @@
-package self
-
-import (
-	"github.com/devsy-org/devsy/cmd/flags"
-	"github.com/spf13/cobra"
-)
-
-// NewSelfCmd builds the 'devsy self' parent command for managing the devsy CLI itself.
-func NewSelfCmd(globalFlags *flags.GlobalFlags) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "self",
-		Short: "Manage the CLI itself",
-	}
-	cmd.AddCommand(NewUpdateCmd())
-	return cmd
-}
